Move closing of results window out of hook
