+++ stockflux-launcher/src/free-text-search/search-result/SearchResultsWindowPosition.js
@@ -1,70 +1,17 @@
 import { Constants } from 'openfin-react-hooks';
 import ReactDOMServer from 'react-dom/server';
+import getWindowPosition from "./SearchResults.positioner";
 
 const RESULTS_WINDOW_NAME = 'child-window-search-results';
-const DEFAULT_LAUNCHER_SIZE = 50;
-const DEFAULT_SEARCH_RESULTS_SIZE = 400;
 
-const getRect = ref => {
-  if (ref.current) {
-    return ref.current.getBoundingClientRect();
-  }
-};
-
-const getResultsWindowPosition = (
-  searchButtonRef,
-  searchInputRef,
-  dockedTo,
-  windowBounds
-) => {
-  const searchButtonRect = getRect(searchButtonRef);
-  switch (dockedTo) {
-    case Constants.ScreenEdge.TOP:
-      const searchInputRect = getRect(searchInputRef);
-      return {
-        defaultTop: DEFAULT_LAUNCHER_SIZE,
-        defaultLeft: parseInt(searchInputRect.left),
-        defaultWidth: parseInt(searchInputRect.width),
-        defaultHeight: DEFAULT_SEARCH_RESULTS_SIZE
-      };
-    case Constants.ScreenEdge.LEFT:
-      return {
-        defaultTop: parseInt(searchButtonRect.top),
-        defaultLeft: DEFAULT_LAUNCHER_SIZE,
-        defaultWidth: DEFAULT_SEARCH_RESULTS_SIZE,
-        defaultHeight: parseInt(searchButtonRect.bottom)
-      };
-    case Constants.ScreenEdge.RIGHT:
-      return {
-        defaultTop: parseInt(searchButtonRect.top),
-        defaultLeft: windowBounds.left - DEFAULT_SEARCH_RESULTS_SIZE,
-        defaultWidth: DEFAULT_SEARCH_RESULTS_SIZE,
-        defaultHeight: parseInt(searchButtonRect.bottom)
-      };
-    default:
-      return {
-        defaultTop: 0,
-        defaultLeft: 0,
-        defaultWidth: DEFAULT_SEARCH_RESULTS_SIZE,
-        defaultHeight: DEFAULT_SEARCH_RESULTS_SIZE,
-      };
-      break;
-  }
-};
-
-const spawnWindow = async (
+export const createWindow = async (
   searchButtonRef,
   searchInputRef,
   dockedTo,
   windowBounds,
-  setResultsWindow
 ) => {
-  const { defaultTop, defaultLeft, defaultWidth, defaultHeight } = getResultsWindowPosition(
-    searchButtonRef,
-    searchInputRef,
-    dockedTo,
-    windowBounds
-  );
+  const { defaultTop, defaultLeft, defaultWidth, defaultHeight } = getWindowPosition(
+    searchButtonRef, searchInputRef, dockedTo, windowBounds);
 
   const childWindow = {
     name: RESULTS_WINDOW_NAME,
@@ -81,36 +28,11 @@ const spawnWindow = async (
     waitForPageLoad: true,
     alwaysOnTop: true
   };
-  await window.fin.Window.create(childWindow).then(win => {
-    setResultsWindow(win);
-    if (
-      dockedTo === Constants.ScreenEdge.LEFT ||
-      dockedTo === Constants.ScreenEdge.RIGHT
-    ) {
-      win
-        .getWebWindow()
-        .document.getElementById('searchbar-container').hidden = false;
+  return window.fin.Window.create(childWindow).then(win => {
+    if (dockedTo === Constants.ScreenEdge.LEFT || dockedTo === Constants.ScreenEdge.RIGHT) {
+      win.getWebWindow().document.getElementById('searchbar-container').hidden = false;
     }
-  });
-};
-
-export const createWindow = async (
-  searchButtonRef,
-  searchInputRef,
-  dockedTo,
-  windowBounds,
-  setResultsWindow
-) => {
-  window.fin.desktop.System.getAllWindows(function(windowInfoList) {
-    if (windowInfoList[0].childWindows.length === 0) {
-      spawnWindow(
-        searchButtonRef,
-        searchInputRef,
-        dockedTo,
-        windowBounds,
-        setResultsWindow
-      );
-    }
+    return win;
   });
 };
